+++ SCP330Behavior.cs
@@ -37,6 +37,7 @@ namespace SCP956
                 noHands = true;
                 HUDManager.Instance.UIAudio.PlayOneShot(BoneCracksfx, 1f);
                 HUDManager.Instance.DisplayTip("Took too much candy", "You feel a sharp pain where your hands should be. They've been severed by an unknown force.");
+                localPlayer.JumpToFearLevel(1f, true);
                 // TODO: Make it so the players hands are no longer visible
 
                 StatusEffectController.Instance.DamagePlayerOverTime(5, 2, true);
@@ -50,7 +51,7 @@ namespace SCP956
             logger.LogDebug("Got Candy");
             candy.spawnPrefab.GetComponent<CandyBehavior>().pinataCandy = false;
             logger.LogDebug("Set pinataCandy to false");
-            int scrapValue = (int)UnityEngine.Random.Range(config9561MinValue.Value, config9561MaxValue.Value * RoundManager.Instance.scrapValueMultiplier);
+            int scrapValue = (int)UnityEngine.Random.Range(configCandyMinValue.Value, configCandyMaxValue.Value * RoundManager.Instance.scrapValueMultiplier);
             logger.LogDebug("Got scrapValue");
             NetworkHandler.Instance.SpawnItemServerRpc(localPlayer.actualClientId, candy.itemName, scrapValue, transform.position, Quaternion.identity, false, true);
             logger.LogDebug("Spawned candy");